+++ modules/gallery/controllers/AlbenController.php
@@ -5,7 +5,7 @@ namespace gallery\controllers;
 use \galleryadmin\models\Album;
 use \galleryadmin\models\Cat;
 
-class AlbenController extends \luya\base\PageController
+class AlbenController extends \luya\base\Controller
 {
     public function actionIndex($catId)
     {