+++ src/base/PageController.php
@@ -2,6 +2,9 @@
 
 namespace luya\base;
 
+throw new Exception("PageController is not in use anymore. Extend your controller from '\luya\baseController'.");
+
+/*
 class PageController extends \luya\base\Controller implements \yii\base\ViewContextInterface
 {
     public $pageTitle = null;
@@ -10,3 +13,4 @@ class PageController extends \luya\base\Controller implements \yii\base\ViewCont
         'pageTitle',
     ];
 }
+*/